api: add the ability to remotely shutoff clients
This is known as the "client kill" mechanism. There are two different
paths, bothing leading to the client being disabled with no network
connection to the SaaS for some period of time:
1. The client receives an unauthenticated response.
2. The client receives a specific runtime generic kill update which can
   be more targeted.

The client will persist a file with a wakeup time after which the client
will come back online and attempt to contact the server again.

Fixes BIT-3577

Signed-off-by: Matt Klein <mklein@bitdrift.io>