client stats: rewrite aggregation to support multiple windows (#45)
This is a substantial rewrite of the client stats local aggregation
functionality. The code now maintains up to N files, each of a maximum
aggregation duration of M minutes. This makes sure that during periods
of offline connecticity we maintain discrete metric windows. Further,
when devices are offline for a long time and then come online again,
new metrics will not get mixed with old metrics.

Fixes BIT-3953
Fixes BIT-2850

Signed-off-by: Matt Klein <mklein@bitdrift.io>