track total sessions seen within workflow engine (#174)
Take two at tracking the total number of sessions. Instead of the
previous attempt that tracked it as part of the session strategy, this
tracks the number of sessions seen by the workflow engine. This is less
fool-proof in that other sessions might have been observed, but is
easier to implement and should be consistent with the number of possible
sessions seen by the workflow engine