improved runtime/workflow crash loop detection
Now we store the nonce of the last accepted/cached configuration. If
we think we have entered a crash loop, we refuse to receive new
configuration unless the nonce has changed. In order to support this
we no longer clean the cached files when we hit the retry count. The
downside of this is that if a client is put into the killed state it
would likely come back online and refuse to accept new config if the
config hasn't changed. To handle this case we clear cached config
when the client enters the killed state.

Fixes BIT-6168

Signed-off-by: Matt Klein <mklein@bitdrift.io>