API shell for feature flags (#246)
This PR connects the feature flags to the running system. The current FeatureFlags is part of the Logger, and the previous FeatureFlags is sent to the crash handler, where it gets added to any reports being sent.

Note that this behavior is not very robust (if we have multiple reports to process, we don't know which one these flags should be associated with), but it is a starting point.

Flags are sent via the async log buffer so that they can be sent before the full system is operational.

FeatureFlagsManager is a temporary object that handles moving current flags to previous flags, and managing copies of each so that they can be sent to their respective destinations.

This is enough to get us out the gates, but we'll need to revisit this to make sure we can handle all edge cases.