hyper network: fix stream closure and handle multiple streams (#286)
The way the hyper network was originally written it didn't really handle
stream closure or retries. This changes it so that we correctly handle
multiple streams and avoids panicing in cases where the stream is being
closed due to remote disconnect.

This is mainly used for the logger-cli and in tests which means that
this is not addressing a live issue.