ring buffer: store logs as proto instead of fb (#297)
For logs with many fields, the fact that fb is padded and doesn't
use varint encoding leads to substantial overhead. This removes
that overhead. As part of this change I went a bit over the top
and also implemented a custom encoder for the Log proto which
allows for zero copy into the ring buffer. This is better than the
previous fb implementation which encoded and then did a memcpy into
the buffer.

Signed-off-by: Matt Klein <mklein@bitdrift.io>